LibSqueezeSeg data inputs

+++ sharedlib/libsqueeze_seg_v2/src/libsqueeze_seg.h
@@ -5,6 +5,6 @@ void
 initialize_python_context();
 
 int*
-libsqueeze_seg_process_point_cloud(int vertical_resolution, int shots_to_squeeze, float* point_cloud, double timestamp);
+libsqueeze_seg_process_point_cloud(int vertical_resolution, int shots_to_squeeze, double* point_cloud, double timestamp);
 
 #endif
